refactor: split analysis/subagents/verify.py into focused modules
